+++ coverage/misc.py
@@ -9,7 +9,9 @@ import inspect
 import locale
 import os
 import os.path
+import random
 import re
+import socket
 import sys
 import types
 
@@ -173,6 +175,18 @@ def output_encoding(outfile=None):
         locale.getpreferredencoding()
     )
     return encoding
+
+
+def filename_suffix(suffix):
+    """Compute a filename suffix for a data file."""
+    if suffix is True:
+        # If data_suffix was a simple true value, then make a suffix with
+        # plenty of distinguishing information.  We do this here in
+        # `save()` at the last minute so that the pid will be correct even
+        # if the process forks.
+        dice = random.Random(os.urandom(8)).randint(0, 999999)
+        suffix = "%s.%s.%06d" % (socket.gethostname(), os.getpid(), dice)
+    return suffix
 
 
 class Hasher(object):